+++ RayTracer/common.hpp
@@ -8,6 +8,7 @@
 #include <algorithm>
 #include <vector>
 #include <tuple>
+#include <stdexcept>
 
 
 #define EPSILON 1e-6
@@ -17,9 +18,9 @@
 #define CPP_IS_FUCKING_RETARDED(x) inline x& operator =(const x& value) noexcept { return this == &value ? *this : *new(this)x(value); }
 
 
-constexpr bool solve_quadratic(const double& a, const double& b, const double& c, double* __restrict const x0, double* __restrict const x1)
+constexpr bool solve_quadratic(const float& a, const float& b, const float& c, float* __restrict const x0, float* __restrict const x1)
 {
-    const double discr = b * b - 4 * a * c;
+    const float discr = b * b - 4 * a * c;
 
     if (discr < 0)
         return false;
@@ -27,7 +28,7 @@ constexpr bool solve_quadratic(const double& a, const double& b, const double& c
         *x0 = *x1 = -.5 * b / a;
     else
     {
-        const double q = (b > 0) ? -.5 * (b + std::sqrt(discr)) : -.5 * (b - std::sqrt(discr));
+        const float q = (b > 0) ? -.5 * (b + std::sqrt(discr)) : -.5 * (b - std::sqrt(discr));
 
         *x0 = q / a;
         *x1 = c / q;
@@ -35,7 +36,7 @@ constexpr bool solve_quadratic(const double& a, const double& b, const double& c
 
     if (x0 > x1)
     {
-        double tmp = *x0;
+        float tmp = *x0;
 
         *x0 = *x1;
         *x1 = tmp;
